+++ roboplan_ros_cpp/CMakeLists.txt
@@ -70,7 +70,7 @@ nanobind_add_module(
   STABLE_ABI
   NB_STATIC
   NB_SUPPRESS_WARNINGS
-  src/bindings.cpp
+  src/roboplan_ros_cpp_bindings.cpp
 )
 target_include_directories(bindings PUBLIC
     $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
